Guys,
Fixed a wee bug and the house caved in. Just about every driver was 
broken in a little evil way and broke the interface conventions in the 
following: 
 
1. The drivers did not correctly update the reference time variable 
pp->lastref. Some left it untouched, some broke the rule that the 
reference time must not be later than the receive timestamp. I added 
pp->lastref = pp->lastrec just before the refclock_receive() call, which 
should be correct in most cases. For those drivers that use 
refclock_process(), coders might want to move the line just before that 
call. 
 
2. There were two variables pp->usec and pp->msec for microseconds and 
milliseconds, respectively, of the second. There was no provision for 
nanoseconds. Some drivers used both of these, which was not handled 
correctly in the interface. I removed both and replaced with pp->nsec 
for nanoseconds and recoded various drivers to do that correctly. 
 
3. Some drivers did not zero the fraction part of the l_fp offset 
variable computed by clocktime(). This was the first thing I noticed in 
testing. The result in some cases was garbage when computing the 
difference between clock time and system time. 
 
4. The SAMPLE macro in ntp_refclock.h and the refclock_sample() routine 
was defective. If the median filter ever filled up, the latter would 
consider it empty and toss the entire filter contents. The intent when 
the filter fills up is that old samples are overwritten and the filter 
retains its maximum size untl refclock_receive() empties it. 
 
5. Small changes were made to interface routines in order to improve and 
simplify the debug trace. 
 
While I carefully checked all the drivers and the changes are minor, I 
am able to test only a few, including the audio drivers, PSTI, 
Spectracom, Arbiter and PPS, which do just fine. I have done nothing for 
the parse drivers, which may or may not need the same updates. I was 
able to confirm the Solaris, SunOS, FreeBSD and Linux builds work. Later 
for HPUX and Alpha. 
 
Aside: In experimenting with the -N high option, which is necessary for 
the WWV audio driver, I discovered some little evils. First, the serial 
port jitter is terrible with my Ultra 30 - simething like 10 ms - and 
the latency surprisingly large, like 10 ms. Checking timestamps, I 
discovered without -N the CR and LF have the same timestamp, while with 
-N they have separate and much more precise timestamps. I also 
discovered for whatever reason the monitor function of the audio codec 
doesn't work in the Ultra 30, but works just fine using the same code in 
the Blade 1000, both Solariba 8. 

bk: 3c23fce6nzgEZzf0wMIaQZAbssPy2w